Move userEmailMap into admin actions
